docker/requirements.llm-provider.txt
# LLM Provider - Requirements
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
aiohttp>=3.9.0
httpx>=0.26.0
redis>=5.0.0
pydantic>=2.5.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
python-json-logger>=2.0.0
tiktoken>=0.5.0
tenacity>=8.2.0
